Fixed types.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@10628 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_vmd/mod_vmd.cpp
@@ -218,8 +218,8 @@ static switch_bool_t process_data(vmd_session_info_t *vmd_info,
     len = frame->samples * sizeof(int16_t);
     data = (int16_t *)frame->data;
     
-    for(max = abs(data[0]), i = 1; i < frame->samples; i++){
-		if(abs(data[i]) > max) max = abs(data[i]);
+    for(max = (int16_t)abs(data[0]), i = 1; i < frame->samples; i++){
+		if((int16_t)abs(data[i]) > max) max = (int16_t)abs(data[i]);
     }
 
 /*
